+++ packages/mk/src/eventHub/index.ts
@@ -1,0 +1,1 @@
+export * from './EventHub';
