+++ packages/mk/src/makeContainer/README.md
@@ -1,0 +1,16 @@
+# makeContainer
+
+Utility for building service containers with lazy or eager resolution.
+
+## Usage
+
+```ts
+import { makeContainer } from '@alexkunin/mk';
+
+const container = makeContainer(builder => builder
+  .lazy('a', () => ({ value: 'A' }))
+  .eager('b', ({ a }) => ({ uses: a.value }))
+);
+
+console.log(container.b.uses); // "A"
+```
