+++ packages/mk/src/makeContainer/index.ts
@@ -1,2 +1,1 @@
-export * from './eventHub';
 export * from './makeContainer';
