+++ packages/mrk/src/makeContext/README.md
@@ -1,0 +1,20 @@
+# makeContext
+
+Creates a React context with lazy initialization.
+
+## Usage
+
+```tsx
+import { makeContext } from '@alexkunin/mrk';
+
+const [ Provider, useCtx ] = makeContext(() => ({ value: 'foo' }));
+
+<Provider>
+  <Child />
+</Provider>
+
+function Child() {
+  const ctx = useCtx();
+  return <span>{ctx.value}</span>;
+}
+```
