+++ packages/mrk/src/makeEventHub/index.ts
@@ -1,2 +1,1 @@
-export * from './makeContext';
 export * from './makeEventHub';
